# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Simple Inventory API",
    "description": "This is a simple API",
    "contact": {
      "email": "you@your-company.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Localhost"
    },
    {
      "url": "https://virtserver.swaggerhub.com/hwupathum/media-server/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "tags": [
    {
      "name": "auth",
      "description": "Used for login, sign up"
    },
    {
      "name": "user",
      "description": "Used for logged in functions"
    },
    {
      "name": "file",
      "description": "Used for file handling"
    },
    {
      "name": "folder",
      "description": "Used for folder handling"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "User login",
        "operationId": "userLogin",
        "requestBody": {
          "description": "Email and password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success / logical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "User register",
        "operationId": "userRegister",
        "requestBody": {
          "description": "Register",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success / logical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "User logout",
        "operationId": "userLogout",
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Need login to fulfill request"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/file": {
      "post": {
        "tags": [
          "file"
        ],
        "summary": "Upload a file",
        "operationId": "fileUpload",
        "requestBody": {
          "description": "Add file using base64 with metadata as json",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadingFile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File Uploaded / logical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Need login to fulfill request"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "file"
        ],
        "summary": "Search files",
        "operationId": "fileSearch",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "schema": {
              "type": "string"
            },
            "description": "File Path",
            "example": "/"
          }
        ],
        "responses": {
          "200": {
            "description": "Files Found / logical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewingFiles"
                }
              }
            }
          },
          "401": {
            "description": "Need login to fulfill request"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/file/{fileId}": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "View a file",
        "operationId": "fileView",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "File Id",
            "example": "9bf5e960-096a-11eb-99a5-b5fd06f70125"
          }
        ],
        "responses": {
          "200": {
            "description": "File View / Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ViewingFile"
                    },
                    {
                      "$ref": "#/components/schemas/Message"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Need login to fulfill request"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "file"
        ],
        "summary": "Delete a file",
        "operationId": "fileDelete",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "File Id",
            "example": "9bf5e960-096a-11eb-99a5-b5fd06f70125"
          }
        ],
        "responses": {
          "200": {
            "description": "File Delete / Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Need login to fulfill request"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/folder": {
      "post": {
        "tags": [
          "folder"
        ],
        "summary": "Create a new folder",
        "operationId": "folderCreate",
        "requestBody": {
          "description": "Folder name with path",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatingFolder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder Created / logical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Need login to fulfill request"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "folder"
        ],
        "summary": "View a folder",
        "operationId": "folderView",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "schema": {
              "type": "string"
            },
            "description": "File Path",
            "example": "/New Folder/"
          }
        ],
        "responses": {
          "200": {
            "description": "Folder View / logical error",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ViewingFolder"
                    },
                    {
                      "$ref": "#/components/schemas/Message"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Need login to fulfill request"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "folder"
        ],
        "summary": "Delete a folder",
        "operationId": "folderDelete",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "schema": {
              "type": "string"
            },
            "description": "File Path",
            "example": "/New Folder/"
          }
        ],
        "responses": {
          "200": {
            "description": "Folder Deleted / logical error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "description": "Need login to fulfill request"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Auth": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "abc@def.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "12345678"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example" : "false"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "password2"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minimum": 2,
            "maximum": 12,
            "example": "John Dove"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "abc@def.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "12345678"
          },
          "password2": {
            "type": "string",
            "format": "password",
            "example": "12345678"
          }
        }
      },
      "File": {
        "type": "object",
        "required": [
          "id",
          "fileName",
          "size",
          "data",
          "ownerId",
          "uploadDate"
        ],
        "properties": {
          "id": {
            "type": "string",
            "maximum": 255,
            "example": "f12fa660-0969-11eb-8225-d5d55e0400df"
          },
          "fileName": {
            "type": "string",
            "maximum": 255,
            "example": "test_file.txt"
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "example": 1024
          },
          "data": {
            "type": "string",
            "format": "byte",
            "maximum": 6e+6,
            "example": "SGVsbG8gV29ybGQ="
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time",
            "example": "2017-07-21T17:32:28.000+00:00"
          },
          "path": {
            "type": "string",
            "example": "/folder/sub_folder/"
          },
          "ownerId": {
            "type": "string"
          }
        }
      },
      "UploadingFile": {
        "type": "object",
        "required": [
          "fileName",
          "size",
          "data"
        ],
        "properties": {
          "fileName": {
            "type": "string",
            "maximum": 255,
            "example": "test_file.txt"
          },
          "size": {
            "type": "integer",
            "example": 1024
          },
          "data": {
            "type": "string",
            "format": "byte",
            "maximum": 6e+6,
            "example": "SGVsbG8gV29ybGQ="
          },
          "path": {
            "type": "string",
            "example": "/folder/sub_folder/"
          }
        }
      },
      "ViewingFile": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example" : "false"
          },
          "file": {
            "$ref": "#/components/schemas/File"
          }
        }
      },
      "ViewingFiles": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example" : "false"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/File"
            }
          }
        }
      },
      "CreatingFolder": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "example": "/New Folder/"
          }
        }
      },
      "Folder": {
        "type": "object",
        "required": [
          "id",
          "ownerId",
          "path"
        ],
        "properties": {
          "id": {
            "type": "string",
            "maximum": 255,
            "example": "f12fa660-0969-11eb-8225-d5d55e0400df"
          },
          "path": {
            "type": "string",
            "example": "/folder/sub_folder/"
          },
          "ownerId": {
            "type": "string"
          }
        }
      },
      "ViewingFolder": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example" : "false"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/File"
            }
          },
          "folders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Folder"
            }
          }
        }
      }
    }
  }
}